requirements.txt
Flask==2.3.2
twilio==8.1.0
python-dotenv==1.0.0
requests==2.31.0
qstash==2.0.5